# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Majed Sultan Al-Subaie Law Firm ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /moj-directory/search:
    get:
      operationId: searchMojDirectory
      tags: [mojDirectory]
      summary: Search MoJ directory by court name
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MojDirectorySearchResult"
  /moj-directory/upload:
    post:
      operationId: uploadMojDirectory
      tags: [mojDirectory]
      summary: Upload MoJ directory PDF (SYSTEM_MANAGER only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MojUploadResult"
        "403":
          description: Forbidden

  # ─── PROFILE (self-service for any authenticated user) ──────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks sorted by due date ascending
      parameters:
        - name: timeframe
          in: query
          required: false
          schema:
            type: string
            enum: [today, week, month]
        - name: show_all
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: show_deleted
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a manual task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (status, title, due date...)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task (creator or SYSTEM_MANAGER)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
        "404":
          description: Not found

  # ─── AI LEGAL ASSISTANT ─────────────────────────────────────────────────────
...
  /ai/ask:
    post:
      operationId: askLegalAssistant
      tags: [ai]
      summary: Ask the AI legal assistant a question (RAG)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskRequest"
      responses:
        "200":
          description: Answer with sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AskResponse"
        "400":
          description: Invalid request
        "404":
          description: No indexed documents
        "503":
          description: AI service not configured
...
  /ai/conversations/{id}/ask:
    post:
      operationId: askInConversation
      tags: [ai]
      summary: Ask a question within a conversation (persists both user msg and answer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskRequest"
      responses:
        "200":
          description: Answer with sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AskResponse"
        "404":
          description: Conversation not found
        "503":
          description: AI service not configured
components:
  schemas:
...
    MojDirectoryEntry:
      type: object
      required: [id, courtName, emailAddress]
      properties:
        id:
          type: integer
        courtName:
          type: string
        emailAddress:
          type: string
    MojDirectorySearchResult:
      type: object
      required: [results, total]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/MojDirectoryEntry"
        total:
          type: integer
    MojUploadResult:
      type: object
      required: [inserted, message]
      properties:
        inserted:
          type: integer
        message:
          type: string

    # ─── NOTIFICATION ─────────────────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, title, taskType, status, dueDate, assignedById, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        taskType:
          type: string
          enum: [MANUAL, HEARING_AUTO]
        status:
          type: string
          enum: [PENDING, COMPLETED]
        dueDate:
          type: string
          format: date-time
        assignedById:
          type: integer
        assignedByName:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        relatedHearingId:
          type: ["integer", "null"]
        caseId:
          type: ["integer", "null"]
        caseName:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: ["string", "null"]
          format: date-time
    CreateTaskInput:
      type: object
      required: [title, dueDate]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        dueDate:
          type: string
          format: date-time
        assignedToId:
          type: ["integer", "null"]
        caseId:
          type: ["integer", "null"]
        linkUrl:
          type: ["string", "null"]
    UpdateTaskInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        dueDate:
          type: string
          format: date-time
        assignedToId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [PENDING, COMPLETED]
        linkUrl:
          type: ["string", "null"]
...
    AskRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 4000
    AskSource:
      type: object
      required: [id, content, documentTitle]
      properties:
        id:
          type: string
        content:
          type: string
        documentTitle:
          type: string
    AskResponse:
      type: object
      required: [answer, sources]
      properties:
        answer:
          type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/AskSource"
...